# service-java/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /reservations/{reservation_id}:
    put:
      tags:
      - reservation-controller
      operationId: updateReservation
      parameters:
      - name: reservation_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReservationRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Reservation'
    delete:
      tags:
      - reservation-controller
      operationId: deleteReservation
      parameters:
      - name: reservation_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /reservations/:
    post:
      tags:
      - reservation-controller
      operationId: createReservation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Reservation'
  /reservations/users/{userId}:
    get:
      tags:
      - reservation-controller
      operationId: getUserReservations
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
  /reservations/roomTypes:
    get:
      tags:
      - reservation-controller
      operationId: getRoomTypes
      parameters:
      - name: checkinDate
        in: query
        required: false
        schema:
          type: string
          default: 2024-02-15T15:00:38.122Z
      - name: checkoutDate
        in: query
        required: false
        schema:
          type: string
          default: 2024-02-16T15:00:38.122Z
      - name: guestCapacity
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomType'
components:
  schemas:
    UpdateReservationRequest:
      type: object
      properties:
        checkinDate:
          type: string
        checkoutDate:
          type: string
    Reservation:
      type: object
      properties:
        id:
          type: integer
          format: int32
        room:
          $ref: '#/components/schemas/Room'
        checkinDate:
          type: string
          format: date
        checkoutDate:
          type: string
          format: date
        user:
          $ref: '#/components/schemas/User'
    Room:
      type: object
      properties:
        number:
          type: integer
          format: int32
        type:
          $ref: '#/components/schemas/RoomType'
    RoomType:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        guestCapacity:
          type: integer
          format: int32
        price:
          type: number
          format: double
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        mobileNumber:
          type: string
    ReservationRequest:
      type: object
      properties:
        checkinDate:
          type: string
        checkoutDate:
          type: string
        roomType:
          type: string
        user:
          $ref: '#/components/schemas/UserDto'
        rate:
          type: number
          format: double
    UserDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        mobileNumber:
          type: string